[FIX] sale_suggested_price dep on product_main_supplier
