+++ sale_suggested_price/__manifest__.py
@@ -13,6 +13,7 @@
     "license": "AGPL-3",
     "application": False,
     "depends": [
+        "product_main_supplier",
         "account",
         "product",
         "purchase",
