+++ sale_suggested_price/models/product_template.py
@@ -1,6 +1,5 @@
 # Copyright 2020 Coop IT Easy SCRL fs
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
-from datetime import date
 
 from odoo import api, fields, models
 from odoo.exceptions import ValidationError
@@ -102,23 +101,6 @@ class ProductTemplate(models.Model):
                 raise ValidationError(
                     _("No Vendor defined for product '%s'") % product.name
                 )
-
-    def _get_main_supplier_info(self):
-        # fixme this function either returns a supplier or a collection.
-        #  wouldn’t it be more logical to return a supplier or None?
-
-        # supplierinfo w/o date_start come first
-        def sort_date_first(seller):
-            if seller.date_start:
-                return seller.date_start
-            else:
-                return date.max
-
-        suppliers = self.seller_ids.sorted(key=sort_date_first, reverse=True)
-        if suppliers:
-            return suppliers[0]
-        else:
-            return suppliers
 
 
 def round_5c(price):
