+++ sale_suggested_price/views/product_supplierinfo_views.xml
@@ -5,6 +5,7 @@
         <field name="model">product.supplierinfo</field>
         <field name="inherit_id" ref="product.product_supplierinfo_tree_view" />
         <field name="arch" type="xml">
+            <!-- todo : replace price with price ?? -->
             <field name="price" position="replace">
                 <field name="price" />
             </field>
